Checking success of mmap
Until now, there was no check to see if the `mmap' system call was
successful. As a result, when it failed, the user would end up with a
segmentation fault later when the array was to be used (which was very
confusing). Now, there is a check to see if `mmap' succeeded and if it
fails, an error (along with a warning) is printed informing the users.

This issue was actually found when running NoiseChisel on a huge image with
`--minmapsize=0', as a result, thousands of `mmap' files were created and
`mmap' failed. But when setting the value of `--minmapsize' to a small
value like 10000 (in bytes, instead of 0) the error was fixed. Since this
might happen to the users using large images, the warning suggests this as
one possible solution.

I also noticed that the common option `--minmapsize' was missing from the
book. So it is now included.