Check on floating point errors in pixel scale calculation
See the added error description for the full discussion: in short, some
programs include floating point errors in their WCS FITS header keywords
and these can bias the pixel scale measurements (and produce an error
message sometimes). So with this commit, a warning will now be printed to
guide the users on the (possible) problem and how to solve it.

This was brought up in a discussion with Leindert Boogaard.