NoiseChisel can now ignore some qthresh tiles
Until now, NoiseChisel would use all the tiles with a successful
measurement. As a result, if we had a region with a large diffuse signal,
the large diffuse signal would be treated like an elevated background and
effectively detected as sky.

From this commit onwards, NoiseChisel has a `--qthreshtilequant' option to
discard tiles above a given quantile.