Automaticly read number of threads when --numthreads=0
When a later configuration file has a fixed value for `--numthreads', until
now there was no way to over-write it with the maximum number of threads
available to the system (which changes from system to system). Also until
now, a value of `0' was unacceptable for this option.

With this commit, when this option has a value of `0', the maximum number
of threads available to the system will be read at the end of
option-parsing and will be used. This will allow you to over-write a fixed
value in a configuration file.