NoiseChisel frees all major allocated memory
While running NoiseChisel with mmap, I noticed that there were two
`gnuastro/mmap_*' files remaining after NoiseChisel finished. So using
Valgrind (and without using `mmap'), the memory allocations/freeing were
checked and the necessary freeing was done in the proper places. Currently
there are only some small bytes that are not freed in the WCS structure
which I believe is because WCSLIB is not properly freeing.