gal_fits_img_read also reads WCS
Until now the two library functions that were used to read a FITS
image/array wouldn't read the WCS. In the programs, the separate
`gal_wcs_read' function was used to read the WCS structure. However,
ultimately the WCS structure of a FITS extension is part of that extension
(and is also included in the generic data container). Therefore it is
counter intuitive for `gal_fits_img_read' and `gal_fits_img_read_to_type'
to not read the WCS structure.

With this commit, these two FITS image extension reader functions also read
the WCS information and put it in their output data container.

Also, the default `--detgrowquant' was slightly raised to 0.7 (from 0.65)
to be less affected by correlated noise.