Checks for libtool in system added at configure time
The Gnuastro tarball ships with its own version of GNU Libtool (from
Gnulib) that is used for compiling and building the programs. So within the
build environment, libtool is recognized. However, if the user doesn't have
libtool, they will not be alerted and later when they want to use
BuildProgram, they will be confronted with an error and get confused: why
did `make check' pass, but I can't use it?

So now, at configure time, we check if the libtool executable is available
within the search path. If it isn't a warning will be printed and the `make
check' BuildProgram test will get skipped. However, if the user later
installs Libtool, they can use BuildProgram.