BuildProgram's libtool invocation with include-dir
The libtool `.la' file keeps all the information about linking, but not the
headers to parse. So for example, if an old Gnuastro was installed in a
system-wide directory, but a user installs a more recent version of
Gnuastro in their `~/.local' directory (and the functions don't match),
then the headers will be from the old version and the linked libraries will
be from the new version.

With this commit an `-I INCLUDEDIR' is added to the libtool invocation so
if the user doesn't give any other directory, by default it looks into the
header files of the same version as the linked libraries.