Configure checks for libcurl for CFITSIO 3.42
Since version 3.42 (most recent as of this commit, released in August
2017), CFITSIO will link with libcurl also for using HTTPS addresses. If
libcurl isn't installed, then CFITSIO will just ignore it. So, on a system
that has libcurl, Gnuastro's configure script would fail (because CFITSIO
needs to link with libcurl).

With this commit, a check for libcurl is added before checking CFITSIO. If
it exists, then it will be linked with the CFITSIO check and if doesn't,
then CFITSIO won't need it either and we can continue.

This fixes bug #52152.