BuildProgram can use a different Libtool .la file
Until now, BuildProgram could only use the installed `libgnuastro.la'
file. This caused a failure when Gnuastro wasn't already installed on the
system and the user ran `make check'. However, this test `PASS'ed, because
there was no check to see if `libtool' fails in BuildProgram!

With this commit, both these issues are fixed: if `libtool' fails (for any
reason, including not finding the `.la' file), it will abort with a message
and it is now possible to identify a different `.la' file to use instead of
the default one (with the `--la' option).