Cosmology functions moved into the library
Until now, Gnuastro's cosmology functions were only available with the
CosmicCalculator program. But these functions can be useful in many other
contexts (other future Gnuastro programs or for users of the library).

With this commit, these functions are now defined in the library and
available for easy usage.

This was done with the help of Boud Roukema.

This completes task #14665.