2D curvature doc: deanthropomorphisation, s/sphericalplane/sphereandplane/
Here are some further edits to the 2D curvature section of gnuastro.texi,
together with a name change from doc/plotsrc/tex/sphericalplane.tex to
doc/plotsrc/tex/sphereandplane.tex, because spherical plane is an oxymoron
in this context. [The equator of the 3-sphere is a flat 2-sphere, and the
equator of the 2-sphere is a straight circle. But that's another topic...]
I reverted from friend/she/her to creature/it/its; I clarified that the
matching of the spacetime unit conversion value to the speed of light in a
vacuum is a solid experimental result, not just a coincidence; and I
reformatted paragraphs to the number of columns that my emacs editor
happened to choose automatically.