Mean and Median filtering added to Arithmetic program
The Arithmetic program (not library currently) now has two new operators:
`filter-median' and `filter-mean'. Currently they are implemented as a
function directly in `bin/arithmetic/arithmetic.c', but we will later move
it to the libraries (either as a separate library in `gnuastro/filter.h' or
as part of an existing header.