Edits to 2D curved distance section
The recent corrections in the "Distance on a 2D curved space" were slightly
modified to be more clear. In particular these two sections were removed in
the first paragraph. Also, a short notice was added to explain that this
section is not intended to be a mathematically thorough review and the
first paragraph was broken into two.

"We might try to do that by embedding the 3D space in a
flat@footnote{`Flat' means in this context that the Pythagorean theorem is
true for every triangle in a space.} 4D space, but that would require good
intuition of the 4D space."

and

"Curvature is something to think of mathematically as intrinsic to a space
itself, but to make things easier, we will do what mathematicians often do,
and think of our 2D surface (we can also call it a -space) embedded in a
flat 3D space."

The first was removed because it ultimately doesn't fully describe how to
visualize in 4D, so simply saying visualizing 4D is "difficult" is
sufficient. Going into too much detail in the introduction can make it
harder for the reader to get started.

The second was applied in a more simpler way by changing "2D surface" to
"2D surface/space" for the same reason as above. We don't want to get the
reader too stuck in details before we have actually done any analysis.

Finally, the paper that Boud introduced in the previous commit message on
the similarity of the speed of light and gravitational waves is now also
put in the footnotes.