Physical nature agnostic WCS columns in MakeCatalog
To generalize the use of MakeCatalog, it now defines the generic `--w1' and
`--w2' instead of the specific `--ra' and `--dec'. The latter two cases can
still be used however. They are each just aliases for one of the former two
cases depending on the input dataset's `CTYPE' keywords. The other
WCS-related options were also renamed to allow this general new naming
convention.