Avoid centering warning by Texinfo
In the "Extending distance concepts to 3D", there were multiple instances
of `@dispmath' in the middle of a paragraph. This caused a warning by
Texinfo on the fact that `center' must be at the start of the line. To
correct it, all `@dispmath's in this section now have their own paragraph
(one empty line before and after them).