Upper-limit crash fixed on label in whole dimension
When a given label in a dataset extends over a whole dimension, the random
number generator would crash until now. The reason was that the size of the
label along that dimension would be the same as the input's size,
therefore, the difference minus one would be the largest possible unsigned
type, which is much larger than the expected range of the integer random
number generator.

To fix the problem, when a label extends over a whole dimension, the value
is now be fixed to zero (to only allow variation along other dimensions).

This fixes bug #52281.