WCS conversion functions using gal_data_t
Until now, the two WCS conversion functions (`gal_wcs_world_to_img' and
`gal_wcs_img_to_world') would explicitly assume RA and Dec and work based
on input arrays (so for example it was also necessary to give the number of
elements and etc). In other words, they weren't using the advantages of
`gal_data_t'. With this commit, they have been modified to accept
`gal_data_t' as input for the input coordinates.

The calls to these functions in Crop, MakeCatalog and MakeProfiles have
also been modified to work with the new API.