Specific calculations in CosmicCalculator
Until now, only two specific calculations could be done in CosmicCalculator
and they were only usable individually. But following the model of
Gnuastro's Statistics program, it is now possible to ask for any of the
calculations in any order and they will be printed in a single line (with
one space character between them).

Also, the option grouping enum has been moved to `ui.h' (from `ui.c') and
their prefix was changed to `UI_GROUP' for the CosmicCalculator and
Statistics programs. In the next commit, this will be done for all the
programs. In the old system it wasn't immediately obvious where they are
defined since they didn't follow the general naming convention.