Option group identifiers corrected to UI_GROUP
The identifiers of option groups in all the different programs started with
`ARGS_GROUP' and the enum defining them was in `ui.c'. This was contrary to
the naming convention in Gnuastro and so made it hard to tell where they
were defined. In the previous commit, this was corrected for the
CosmicCalculator and Statistics programs. In this commit, it is corrected
for all the other programs.

Now all program specific option groups are defined in the program's `ui.h'
and they start with `UI_GROUP'.