BuildProgram deletes the compiled program if requested
BuildProgram can also run a program that it builds. This is the default
behavior. Therefore, in some situations, you don't want the actual
binary/compiled program, you just want to see the result of some operation
and you will later delete the compiled file.

With this commit, BuildProgram has a new `--deletecompiled' option. With
this option, when the built program is run (when `--onlybuild' isn't
called), BuildProgram will delete the compiled binary.

Two other minor corrections:

 - The number of threads are irrelevant to BuildProgram, so it is now a
   hidden option for it.

 - The "Single row calculations" option group title in CosmicCalculator was
   changed to "Specific calculation" to fit the terminology of the book and
   code.