Library table writing functions don't delete
Until now the table writing functions had a `dontdelete' argument. This
argument was in charge of letting the function know what to do with the
output file if it already existed. But ultimately, this job (deciding what
to do with the output file) is a choice at another level. It would thus
make it hard to write multiple FITS tables to separate extensions.

So with this commit, the choice of what to do with the output file is left
to the caller of the function. If the output is a FITS file and it already
exits, they will write the output as a new extension and if it is a text
file, it will abort with an error (it is the caller's responsibility to
check before calling this function).

With this change, MakeCatalog can now write its two outputs in separate
extensions of a single FITS file. Also in this commit, the options relating
to output catalog columns were grouped into related contexts to make it
easier to find them.