Range for random sampling in Upper-limit measurements
Until now, MakeCatalog would use the full range of the input dataset to
place a random mapping of the labeled region. This would give a biased
result when the noise properties of the input vary (gradually) along a
dimension.

MakeCatalog now has a new `--uprange' option to allow users to specify a
range of pixels along each dimension of the input. The random samplying
will only be done within the given range of each object's position.