More clear MakeCatalog --help output
The group of options related to settings of upper limit magnitude
measurements had a title called "Upper limit magnitude". This could cause
people to confuse these with output columns (following the recent
classification of columns). So this title is now set to "Upper limit
magnitude settings".

Also, MakeCatalog doesn't currently produce a log file, so it is now set to
a hidden option in MakeCatalog and thus it won't show in the output of
`--help'.

I also noticed that the output of `--cite' for MakeCatalog doesn't have new
line characters on its first two lines. This is also corrected here.