Default random number generator set to gsl_rng_ranlxs1
Until now, the three MakeCatalog, MakeNoise and MakeProfile would use
`gsl_rng_default' as the random number generation algorithm. When
unchanged, this was the `mt19937' algorithm. However, the GSL documentation
suggests that "For the most reliable source of uncorrelated numbers, the
second-generation RANLUX generators have the strongest proof of
randomness". There are three luxary levels, so the default value is set to
the middle one (`ranlxs1').