Range of random samples printed in output catalog
In the previous commits, the `--uprange' option was added to MakeProfiles
to specify a range for random values. But these important numbers weren't
written into the output's comments for the users to be able to
reproduce/understand the results. With this commit, when `--uprange' is
specified, the values given to it are also printed as one comment line in
the output.