Checks for GNU Libtool, allowing libtool or glibtool
Until now, Gnuastro assumed that the `libtool' executable is GNU Libtool,
so it would use its options and etc. But thanks to Alan Lefor, I found out
that on Mac systems, the `libtool' executable may not be GNU Libtool, but
Mac's own implementation of it. Mac's implementation doesn't support long
options or the special behaviors we need in BuildProgram. I then found out
that people commonly call GNU Libtool on a Mac with `glibtool'.

With this commit, at configure time, it will be checked if the `libtool'
executable is GNU or not. If it isn't, or if there isn't any `libtool'
executable at all, it will check for `glibtool'. If `glibtool' also isn't
present, then  In any case, BuildProgram will use the proper executable
name.