NoiseChisel doing final growth with max connectivity
When doing the final growth of true clumps, until now, NoiseChisel used
4-connectivity to expand the clumps. However, earlier, it defined the
detections using 8-connectivity. As a result, in some rare cases where the
detection had an 8-connected region to the main body with no true clump in
it, that region was left unlabeled after the final growth and would somehow
get a wrong label (similar to another object in a completely separate part
of the image).

The main cause for this choice was that we want the initial growth of true
clumps to be based on 4-connectivity to have 8-connected rivers. So, to fix
the problem, With this commit, a `connectivity' argument is given to
`clumps_grow'. Thanks to this argument, now, it is possible to ask the
function to use 8 connectivity when we want to cover the full area (don't
want rivers) and 4-connectivity when we want rivers.

Also, a small sanity check was added to `GAL_DIMENSION_NEIGHBOR_OP' to make
sure the connectivity value is reasonable.

This fixes bug #52327.