NoiseChisel can take image instead of doing convolution
When testing NoiseChisel's parameters for different datasets/analysis, it
is really troubling that usually, the single most time (and CPU) consuming
step is convolution. NoiseChisel now has a new `--convolved' option so the
user can specify the convolved image, thus telling NoiseChisel to read that
image instead of doing convolution.