Wrong option name fixed in NoiseChisel
Initially, I had set the name of the new `--convolvedname' option to
`--convolved'. Later during its development, I changed it to its final name
to be similar to the other file name options. But in a few places (that
were not significant to the build), the old name remained. With this
commit, they are now also set to the final name.