NoiseChisel's --convolvedname option back to --convolved
In a previous commit, I had mistakenly assumed that all the file name
options end in `name'. But in NoiseChisel, this was the standard in
`main.h', not the options. We should probably fix this sometime (so the
names in `main.h' and the option names exactly correspond). So with this
commit, the `name' suffix is removed and the option is only called
`--convolved'.

Also, I noticed that the `--log' option is present in NoiseChisel's
`--help' output while NoiseChisel doesn't produce any logs. So in
`bin/noisechisel/ui.c', `--log' is now set to be a hidden option.