Added French translation of top webpage
This is the French translation of the gnuastro software page that you would
find at www.gnu.org/software/gnuastro. Hopefully this will allow more
French users to be introduced to Gnuastro, but there's still lots of work
to be done before they can use the software without knowledge of English.

French is my second language and there's bound to be a few errors or things
that sound odd so it would be best if a native French speaker could review
the document and correct anything that maybe confusing to the reader.