WCS pixel scale prints warning instead of aborting
When the pixel scale finding algorithm of `gal_wcs_pixel_scale' can't find
the proper order of the pixels, until now it would abort with an error. But
now, it prints a warning to `stderr' instead of aborting. Just to let the
user know that there may be problem. In some cases it may not be a problem
for the user.

This issue was brought up by Alejandro Serrano Borlaff.