SV decomposition order uses maximum
Until now, the SV decomposition that was used to correct the pixel
resolution (for CDELT) was re-ordered only if there was one non-zero
element. But now, it will use the maximum value to allow usage on more
datasets (that don't necessarily have the same resolution, for example
tilted images).

This follows the previous commit, so Alejandro's name has been added to the
`THANKS' file also.