Check if any input tokens given to Arithmetic
Until now, unlike the other programs, the Arithmetic program wouldn't
complain if no tokens was given. Therefore, the first time that it was
necessary to read the tokens, there was nothing to read and Arithmetic
would abort with a segmentation fault. With this commit, a check has been
added to let the user know in this case.

This bug was reported by Lee Spitler.

This fixes bug #52422.