BuildProgram now accepts libtool's --tag option
Raúl Infante Sainz reported that the test for BuildProgram failed on his
system with an error of "link: unable to infer tagged configuration, error:
specify a tag with '--tag'". Libtool uses the `--tag' option when it can't
infer the language automatically.

With this commit this issue is fixed with a new `--tag' option to
BuildProgram. It will be directly passed onto Libtool and is set to `CC'
(for the C programming language), by default.

I also noticed that BuildProgram's configuration file wasn't being
copied/used during `make check'. This is also now corrected.