Simple for loop replaced with while
When looking at the type of the WCS coordinates in MakeCatalog, we need to
parse over the input type values. To do that, we just need to loop until we
reach the end of the line or a `-' character. This was implemented as a
`for' loop.

But Raúl Infante Sainz informed me of the following compiler warning while
building Gnuastro on his system: "explicitly assigning value of variable of
type `char *' to itself". In general, a for loop is not designed for such
parsings. So with this commit, we now do this with a while loop which will
also remove this warning.