NoiseChisel initializing snind when finding true clumps
Until now, when there was no clumps in a tile, we sould only set the
`cltprm->sn' pointer to NULL. We also use the `cltprm->snind' in some
places to see if a check image is necessary. However, when there are no
clumps in a tile, the latter wasn't being initialized. It is now
initialized to NULL along with the former.