Removed semi-colon in macros of authors-cite.h
While trying to find the cause of bug #50957 (`--version' output not
possible on Mac OS X), I noticed that the macros in all the programs'
`author-cite.h' have a semi-colon (;) after them. This is not the cause of
the problem, but can cause other bugs (if the macros are used in the middle
of a string). So they are now removed.