Calloc instead of malloc when reading fits keyword comments
When allocating space for a keyword `gal_fits_key_read_from_ptr' would use
`malloc' to just allocate (and not clear/initialize) the memory. But when
the keyword doesn't exist, the comment's space will also not be filled and
so checking it will use an un-initialized value. So we use `calloc' now to
keep space for the comments and units.

Also, when reading a text table column information, we wouldn't check if
the only character left is a string-null. This could happen when the values
inside braces were like `[,,nan]'.