C++ compiler check for -Qunused-arguments
When checking for PThreads, `ax_pthread.m4' will check the C compiler for
possible options to avoid repetative arguments. This includes
`-Qunused-arguments'. It just does the check on the C compiler, but on some
systems (as reported in bug #52490), the C++ compiler might not recognize
it.

So we are using the GNU Autoconf Archives M4 tests for compile flags to see
if the C++ compiler also recognizes this option. If it doesn't, for the
time being, the C++ test in `make check' won't be run. The ideal way would
be to correct `CXXLINK' (created in the produced `tests/Makefile.in') and
remove this option (which comes in through `LDFLAGS') if it is not
recognized by the C++ compiler.

If using a previous Gnuastro bootstrap, please re-do the bootstrap after
this commit (a new Autoconf Archives is now needed).