Minor corrections in French top page
"Gnuastro" was removed from the top header of the mailing lists in the
French top page, since it would get too long. This made the side menu at
the top of the page too wide.

Also, I noticed that I had missed a `.' at the end of the sentence that I
added mentioning that Lucas is in charge of the French translation. This
has been added.