Text table examples in Invoking Match section of book
I noticed that all the examples in the "Invoking Match" section of the book
used FITS files. This could cause confusion to readers that it only accepts
FITS tables. So now, two of the examples are text files and I have also put
a link in the explanations on the table formats.