gal_match_coordinates returns full permutation
Until now, the match library function (`gal_match_coordinates') would just
return the permutation for those elements that match. Therefore the
permutations did not have the same number of elements as the input and the
library's permutation functions couldn't be used. As a result, it was very
hard to use it.

With this commit, this function returns the full permutation of each
dataset. The permutation will put the non-matching elements in the end of
the list. This actually simplified the Match program also.

Match also prints the number of matches to the standard output if `--quiet'
wasn't called.

With this commit, this function is also fully described in the book.