Matching in 1D and columns specified in one option call
Until now match would only work on 2D positions, but now it also matches 1
dimensional positions. This can be handy for example when you want to match
spectral features.

Until now, it was necessary to call `--ccol1' and `--ccol2' multiple times
to determine the coordinate columns. But this would make it hard to deal
with possibly many dimensions. Because for a default behavior, other calls
to these options would be given in a configuration file. It was also
confusing and buggy to call `--ccol1' and `--ccol2' many times.

So now, the (possibly) multiple column identifiers must be given in a
comma-separated list and one call to these options. With this change, the
dimensionality of the match can easily be determined from the value to the
first column identifier option and it is also possible to have these
options in a configuration file.