Two small/important corrections in Match
When reading the second input table after the match, Match was incorrectly
looking into the HDU of the first input! This is now corrected.

Also, Match was not printing a new line after the total number of matches
when not in `--quiet' mode.

Finally, we were mistaknely comparing the `blow' (an index in second input)
with the first input's number of rows (`ar') in
`match_coordinates_second_in_first'. To speed up things when we read the
end of the second catalog, there is now a simple condition just at the
start of the for loop.