Match can also output the non-matched rows
Until now, the output of Match were only the matching rows. Therfore both
output tables had the same number of rows. But in many contexts (especially
when you want to check the match parameters), it is also necessary to
inspect the rows that don't match.

With this commit the new `--notmatched' option has been added to Match,
with this option, the output tables will contain the rows that don't
match.