Matched output can be a merging of both input columns
When the user wanted specific columns of the two inputs, until now Match
would output the whole of both inputs in an arranged order. So another step
was necessary to keep certain columns for higher level analysis. Match now
has a `--outcols' option which the user can use to specify which columns of
the two inputs they want in the output.

To make this possible, it was necessary to have the number of columns
matched with certain column selection options. Therefore `gal_table_read'
was given a new pointer argument. When this pointer is not NULL, the number
of columns in the output that match a given input are stored in the
respective space.