Mohammad's email address corrected
Until now, I had used my `gnu.org' email address in the manual. But to
unify all my emails (from the universities, GNU and Gmail), its a while
that I am using `akhlaghi.org'. So it would be easier for this email
address to also be present in the manual and in the Git history.

I also generated a new PGP key (with this email address), so that has also
been updated in the release checklist.