Added acknowledgements and better info to --cite output
The introductory information in the output of `--cite' was slightly updated
to be more clear and emphasize on the fact that different programs may have
different papers to cite. Also, a text was added to include in the
acknowledgements section of works that use Gnuastro.

Finally, the list of people to thank have been updated in the
"Acknowledgements" section of the book.