Arithmetic library's flag argument converted to int type
The `flags' arguments to `gal_arithmetic' had an `unsigned char'
type. However, on some LLVM compilers, we got the following warning when
using `va_start': "promotion to 'va_start' has undefined behavior" and
"note: parameter of type 'unsigned char' is declared here" (pointing to the
`flags' argument of `gal_arithmetic').

To avoid this warning and have a clean build, in the arithmetic library,
the flags now have an `int' type.

This warning was first reported by Raúl Infante Sainz.