In configure, current directory removed from PATH
This is a followup of commit eaade0ba09b26fd: "Check for libtool rejects
Gnuastro's own libtool implementation". There, I had used the last argument
to AC_CHECK_PROG incorrectly. So following a similar issue that was
reported by Nicolas Bouché, I am trying to solve the problem more
elegantly.

The main problem is that the PATH variable on some systems might contain
the current directory in the form of `./' or `.'. In such cases, if
Gnuastro's Libtool is already built, it will be found as `.//libtool' or
`./libtool' respectively.

With this commit, a more fundamental fix to the problem is implemented: to
correct `PATH' (only for the configure script) so it doesn't look into the
current directory for any program.