Configuration warning if PATH includes current directory
Having the current directory in PATH can be a major security risk. Since we
do the check here in Gnuastro, it now prints a warning at the end to inform
the users of this risk. It is irrelevant to Gnuastro, but can be useful for
users who have done this without understanding its risks. The major risks
are described in the warning message.