Minor edits in book's configuration options and intro
The last paragraph of the "Gnuastro configure options" section was still
based on the old case where all the programs only worked on images and
MakeProfiles' check was necessary for all the other programs. This not the
case any more, so that statement was corrected.

Also, a small correction was made in the "Science and its tools" section to
be more clear.