No -pthread flag in LDFLAGS
When calling the `AX_PTHREAD' macro in `configure.ac', we were adding
`PTHREAD_CFLAGS' to `LDFLAGS'. This was extra, because the `-pthread' flag
is both for the C compiler (to tell it to prepare for multiple threads) and
the linker. Now, when Gnuastro creates the library, it only calls
`-pthread' twice (instead of three times before). I am still trying to find
where the extra `-pthread' comes from!