Minor edits in book, acknowledgements updated
Some minor corrections and edits were made in the libraries section of the
book to fix a mistake and make the text more clear to read.

Also, I had forgot to thank Alfred M. Szmidt from GNU who kindly helped in
connecting me to the GNU Webmasters and finally the GNU French Translation
Team to make the French page accessible. I also added a "Teams" section to
the `THANKS' file to acknowledge the help from that team.