-Qunused-arguments now dealt with in CXXFLAGS
Until now, we would completely disable the C++ check if it didn't recognize
the `-Qunused-arguments' that `AX_PTHREAD' defines in
`PTHREAD_CFLAGS'. This was because we were putting `PTHREAD_CFLAGS'
directly into `CFLAGS' and `LDFLAGS'. Since the latter is used for both C
and C++, it would cause problems if C++ didn't recognize
`-Qunused-arguments'.

From this commit onward, we only add `PTHREAD_CFLAGS' to `CFLAGS' and
`CXXFLAGS' and leave `LDFLAGS' alone. Also, `CXXFLAGS' is checked during
configuration to see if it contains `-Qunused-arguments'. If so, it will
check if C++ can compile with it or not. If not, this option will be
removed from the final `CXXFLAGS'. So now, even if the C++ compiler doesn't
recognize this option, we don't have a problem.