Updated copyright information
2018 was added as the final copyright year in all the files with a
copyright year. For those that already had a single year, that was changed
to a format of `AAAA-BBBB' where `AAAA' is the year the file was created
and `BBBB' is (currently) 2018. This will allow automatic scripts to update
the last year automatically from the next year.

My email address was also updated to `mohammad@akhlaghi.org' from the old
`akhlaghi@gnu.org'.