Corrected copyright date in configure.ac
In the correction of two commits ago, the dates of the copyright were
corrected to have the file starting date and current date (2018). However,
the current date for `configure.ac' was not corrected! This was because it
already had two years but with a space between the dash and the years! So I
mistakenly just corrected the space and forgot to correct the year. This is
corrected now.