Return value of asprintf now checked
When `asprintf' is unable to allocate the necessary memory, it will return
a value that is less than zero. To avoid hard-to-debug errors, a check was
added to every call of `asprintf' so if this happens, the calling program
aborts.

This check was motivated by "ignoring return value of 'asprintf', declared
with attribute warn_unused_result" compiler warnings that came up on some
systems (debian package builders and also reported by Éric Thiébaut).

This fixes bug #52979.