Lzip's output piped into Tar in book's quick start
Antonio Diaz Diaz kindly noted that for systems where Tar doesn't recognize
`.tar.lz', we can use a pipe instead of two commands. So a footnote was
added to the "Quick start" section of the book to remind the users of this
method. In the main text, the two command format is kept because it is more
clear to the un-experienced eyes.

An experienced person in the Unix-way will never even open this section of
the book, they will have their own methods to uncompress and un-pack a
tarball. But the footnote can be a good reminder for a person who is
reading the manual a second time (and has gained some prior experience in
Unix).