Updated Debian release checklist
The main repository hosting packages in Debian has changed to
`salsa.debian.org', as a result it was necessary to also update the
checklist so it corresponds to the correct location.

Also, I had forgot to give a signature file name for the `-s' option to
`pristine-tar'. To make things clear, I added commands to download the
tarball and its signature file so when we get to this `pristine-tar'
command, everything is ready.