Citation discussed in top webpage, ASCL code included
The top webpage lacked a section to discuss the recommended way of citing
Gnuastro. So a small section was added to the top English page (a
translation request to the GNU French Translators team will be sent after
this commit).

Gnuastro was recently registered on the Astrophysical Source Code Library
(ASCL). Therefore, the ASCL registration id is now also included in the
output of `--cite' (as well as the new section on the top webpage).