Correction in Sience and its tools section of book
A minor (one word) correction was made to the ending of the "Science and
its tools" section of the book: "known" was changed to "existing". This was
because "known" could be interpretted as "true": implying "know"ing
something that is "absolutely true". This is contrary to the spirit of this
section, or science as defined by Popper or Kuhn, were the assumption of an
"absolute truth" is not necessary. "existing" was a good alternative to fix
this issue and does not cause the confusion above.