User -I and -L values in BuildProgram's configuration file
When the user passes special values to the `-I' and `-L' compilation
options through the `CPPFLAGS' and `LDFLAGS' directories at configure time
(for example if GSL is installed in a non-standard place), it is also
necessary to pass those directories to BuildProgram. This was not done
until now, so it would fail at `make check'.

We now read the values to `CPPFLAGS' and `LDFLAGS' at configure time and in
BuildProgram's `Makefile.am', we write these values into a new
configuration file which is in the build directory and will be used for
checks and installed. Therefore, the original BuildProgram configuration
file (in the source directory) is now called `astbuildprog.conf.in'.

This issue was found by Michel Tallon.

This fixes bug #53122.