NEWS file and library soname updated for next release
The NEWS file hadn't been prepared to host updates/changes since the 0.5
release. So a title and empty sections for the next release is now ready to
host all the different major updates.

Also in preparation for the next release, the library's major soname
version was updated to 4. In time many changes will certainly be made to
the library before the next release.