Known issues solution to LD_LIBRARY_PATH re-written in book
This solution was slightly out-dated (more than one test will pass and many
will be skipped). It also wasn't generic enough (describing the possibility
of other installation directories). So it was re-written to adapt to what
happens now.

Also instead of the script writing the `export ...' command into
`~/.bashrc', the line is now written and the user is requested to add it
them selves (how ever they like). This is more clear and informative.