gal_options_check_version also returns version
Until now, we had not considered the fact that with the `-P' option, a user
might want to print the version information also (if the `onlyversion'
option is called). Therefore, in such cases, `arg' (first input to
`gal_options_check_version' was NULL and we would get a segmentation
fault). This has been fixed now and this function also returns an allocated
version string to feed into the printing function.

This fixes bug #53142.