All options checked for printing only when necessary
Until now, the `gal_options_print_state' would first go through all the
options to see how many printing options were requested. It would then
check (and abort) if more than one such option was called. Then it would go
over all the options again and print the options if necessary in the proper
place.

In the meantime, the definition of the `--onlyversion' option was unique in
that its `value' pointer was set to `NULL' instead of pointing to somewhere
in memory (in `gal_options_common_params').

As a result, when `--onlyversion' was called, its `set' value will be `1',
but its pointer was still NULL. This caused a crash when we were checking
to see if the user wants to print the option values or not (in the second
round of checks mentioned in the first paragraph above).

Therefore the first fix of this commit is to avoid the second check when no
printing was necessary (to increase efficiency). The main fix was then to
set `onlyversion's `value' pointer definition to a new (but redundant)
`cp->onlyversion' element. Since this structure is initialized to zero in
`main.c', when reading the option values `OPTIONS_UINT8VAL' will read it as
a zero and the main crash above is thus avoided.

This fixes bug #53147.